Config run React in Heroku
