+++ app/controllers/react_app_controller.rb
@@ -1,0 +1,4 @@
+class ReactAppController < ApplicationController
+  def home
+  end
+end
