+++ config/react_app.js
@@ -1,0 +1,3 @@
+module.exports = {
+  api_base_url: "/api/",
+}
